+++ ext/dry_validation_rust/src/lib.rs
@@ -255,6 +255,7 @@ fn init(ruby: &Ruby) -> Result<(), Error> {
     let native: RModule = ruby.eval("Dry::Validation::Rust::Native")?;
     let class = native.define_class("Engine", ruby.class_object())?;
     class.define_singleton_method("new", function!(Engine::new, 1))?;
+    class.define_singleton_method("new_cached", function!(Engine::new_cached, 2))?;
     class.define_method("call", method!(Engine::call, 1))?;
     class.define_method("call_json", method!(Engine::call_json, 1))?;
     class.define_method("register_plugin", method!(Engine::register_plugin, 3))?;